PRACTICA 1/requirements.txt
beautifulsoup4==4.11.1
bs4==0.0.1
certifi==2022.9.24
charset-normalizer==2.1.1
getuseragent==0.0.7
idna==3.4
numpy==1.23.4
pandas==1.5.1
python-dateutil==2.8.2
pytz==2022.6
PyYAML==6.0
requests==2.28.1
six==1.16.0
soupsieve==2.3.2.post1
urllib3==1.26.12
useragent==0.1.1
yml==0.0.1
